requirements.txt
Django
djangorestframework
drf-yasg
numpy
pandas
psutil
psycopg2
psycopg2-binary
requests
django-cors-headers
djoser
